+++ src/DTO/Request/Payment.php
@@ -50,9 +50,9 @@ class Payment
     private $payerNote = null;
 
     /**
-     * @var null
+     * @var string
      */
-    private $description = 'PrestaShop payment';
+    private $description;
 
     /**
      * @return string
@@ -119,7 +119,7 @@ class Payment
     }
 
     /**
-     * @return null
+     * @return string
      */
     public function getDescription()
     {
@@ -127,7 +127,7 @@ class Payment
     }
 
     /**
-     * @param null $description
+     * @param string $description
      */
     public function setDescription($description)
     {