+++ src/Repository/SaferPayOrderRepository.php
@@ -75,4 +75,14 @@ class SaferPayOrderRepository
 
         return Db::getInstance()->executeS($query);
     }
+
+    public function getPaymentBrandBySaferpayOrderId($saferpayOrderId)
+    {
+        $query = new DbQuery();
+        $query->select('`brand`');
+        $query->from('saferpay_assert');
+        $query->where('id_saferpay_order = "' . (int) $saferpayOrderId . '"');
+
+        return Db::getInstance()->getValue($query);
+    }
 }