+++ src/Service/Request/InitializeRequestObjectCreator.php
@@ -30,6 +30,7 @@ use Invertus\SaferPay\Config\SaferPayConfig;
 use Invertus\SaferPay\DTO\Request\RequestHeader;
 use Invertus\SaferPay\DTO\Request\Initialize\InitializeRequest;
 use Invertus\SaferPay\DTO\Request\Payer;
+use PrestaShop\PrestaShop\Adapter\Shop\Context;
 
 if (!defined('_PS_VERSION_')) {
     exit;
@@ -68,8 +69,14 @@ class InitializeRequestObjectCreator
         $totalPrice = (int) (round($totalPrice));
         $payment = $this->requestObjectCreator->createPayment($cart, $totalPrice);
         $payer = new Payer();
+
+        $languageCode = !empty($cart->getAssociatedLanguage()->iso_code)
+            ? $cart->getAssociatedLanguage()->iso_code
+            : 'en';
+
+        $payer->setLanguageCode($languageCode);
         $returnUrl = $this->requestObjectCreator->createReturnUrl($returnUrl);
-        $notification = ($isBusinessLicence && version_compare(Configuration::get(RequestHeader::SPEC_VERSION), '1.35', '<')) ? null : $this->requestObjectCreator->createNotification($customerEmail, $notifyUrl);
+        $notification = $isBusinessLicence ? null : $this->requestObjectCreator->createNotification($customerEmail, $notifyUrl);
         $deliveryAddressForm = $this->requestObjectCreator->createDeliveryAddressForm();
         $configSet = Configuration::get(SaferPayConfig::CONFIGURATION_NAME);
         $cssUrl = Configuration::get(SaferPayConfig::CSS_FILE);