+++ upgrade/install-1.1.8.php
@@ -27,10 +27,18 @@ use Invertus\SaferPay\DTO\Request\RequestHeader;
 if (!defined('_PS_VERSION_')) {
     exit;
 }
-function upgrade_module_1_1_4()
+function upgrade_module_1_1_8(SaferPayOfficial $module)
 {
+    Configuration::updateValue(SaferPayConfig::SAFERPAY_ALLOW_SAFERPAY_SEND_CUSTOMER_MAIL, 1);
+    Configuration::updateValue(
+        SaferPayConfig::SAFERPAY_PAYMENT_DESCRIPTION,
+        SaferPayConfig::SAFERPAY_PAYMENT_DESCRIPTION_DEFAULT_VALUE
+    );
+
     Configuration::updateValue(RequestHeader::SPEC_VERSION, SaferPayConfig::API_VERSION);
     Configuration::updateValue(RequestHeader::SPEC_REFUND_VERSION, SaferPayConfig::API_VERSION);
+
+    $module->registerHook('actionObjectOrderPaymentAddAfter');
 
     return true;
 }